refactorings

+++ jamhh/src/be/ac/ulg/montefiore/run/distributions/PoissonDistribution.java
@@ -74,7 +74,7 @@ public class PoissonDistribution
     // Based on Stirling approximation
     private double logFactorial(int n) {
         if (n >= logFactTable.length) {
-            double r = 1. / (double) n;
+            double r = 1. / n;
 
             return (n + .5) * Math.log(n) - n + C0
                     + r * (C1 + r * r * (C3 + r * r * (C5 + r * r * C7)));
